% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as_degrees}
\alias{as_degrees}
\title{Format radians as degrees}
\usage{
as_degrees(rad, digits = 0)
}
\arguments{
\item{rad}{Values in radians}

\item{digits}{Number of digits to display}
}
\value{
A character vector with degree equivalents of \code{rad}
}
\description{
Utility function for formatting radians as degrees, with control for number
of digits.
}
