% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocdr.R
\name{find_const_pitch}
\alias{find_const_pitch}
\title{Find constant pitch periods within descent}
\usage{
find_const_pitch(data, desc_id, width, fs, max_flat = 10, mae_thr = 0.5
  * pi/180)
}
\description{
Looks for periods in the descent with the lowest mean absolute error (MAE)
from a constant pitch. Finds up to \code{max_flat} periods, ranked by lowest MAE,
excluding any with MAE greater than the \code{mae_thr} threshold.
}
\examples{
desc_samples <- find_desc(prh$depth, prh$pitch, prh$fs) \%>\%
  tidyr::drop_na(desc) \%>\%
  dplyr::group_by(desc) \%>\%
  dplyr::group_modify(~ find_const_pitch(.x, .y, width = prh$fs, fs = prh$fs))
}
