% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_desc.R
\name{find_desc}
\alias{find_desc}
\title{Find steep descent phases at the beginning of dives}
\usage{
find_desc(prh, depth_thr = 5, pitch_thr = -45 * pi/180)
}
\arguments{
\item{prh}{A PRH data frame}

\item{depth_thr}{Minimum depth of a dive}

\item{pitch_thr}{Minimum pitch to count as steep descent (in radians)}
}
\description{
Annotates a PRH data frame with descent phases. Each descent phase begins
with the start of a dive and ends at the beginning of the bottom phase. Only
the part of the descent exceeding the pitch threshold is annotated.
}
