% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{findpeaks}
\alias{findpeaks}
\title{Find peaks in a vector}
\usage{
findpeaks(x, width, thr)
}
\arguments{
\item{x}{Numeric vector}

\item{width}{Size of rolling window}

\item{thr}{Minimum value of x to count as a peak}
}
\value{
Data frame with columns i (index of peaks) and x (value at peak)
}
\description{
Identifies peaks as points where the vector intersects the maximum value in
a window centered on the point
}
