% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit jiggle model}
\usage{
fit_model(prh, n_iter = 100)
}
\arguments{
\item{prh}{PRH data frame with \code{ocdr} and \code{rms_acc} columns}

\item{n_iter}{Number of iterations to try fitting the non-linear model}
}
\value{
PRH data frame with \code{speed} column. Adds an attribute,
\code{speed_model}, a list with elements \code{a}, \code{b}, \code{c1},
\code{c2}, \code{c3}, and \code{predict_speed}. The first five elements
are the coefficients of the model and \code{predict_speed} is a function
that takes an acceleration RMS matrix and returns the predicted speed.
}
\description{
Multiple exponential regression of OCDR on the "jiggle", i.e. root mean
square amplitude of filtered triaxial acceleration.
}
\examples{
prh_ocdr <- get_ocdr(find_desc(prh_expl))
acc_filt <- jgl_filtacc(Araw_expl, 10, 90)
prh_rms <- get_rms_acc(prh_ocdr, acc_filt)
prh_spd <- fit_model(prh_rms)

}
