% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms_acc.R
\name{get_rms_acc}
\alias{get_rms_acc}
\title{Root mean square of acceleration}
\usage{
get_rms_acc(prh, A)
}
\arguments{
\item{prh}{PRH data frame with \code{ocdr} column and \code{binwidth}
attribute.}

\item{A}{Acceleration data frame. Columns \code{Atime} (POSIXct) and
\code{A} (3-column matrix) with an attribute, \code{Afs}, for the sampling
frequency.}
}
\value{
PRH data frame with \code{rms_acc} matrix column.
}
\description{
Calculate root mean square (RMS) of acceleration in bins corresponding to
OCDR bins.
}
\examples{
prh_ocdr <- get_ocdr(find_desc(prh_expl))
acc_filt <- filter_acc(Araw_expl, 10, 90)
get_rms_acc(prh_ocdr, acc_filt)

}
