% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgl_accrms.R
\name{jgl_accrms}
\alias{jgl_accrms}
\title{Root mean square of acceleration}
\usage{
jgl_accrms(prh, A)
}
\arguments{
\item{prh}{PRH data frame with \code{ocdr} column and \code{binwidth}
attribute.}

\item{A}{Acceleration data frame. Columns \code{Atime} (POSIXct) and
\code{A} (3-column matrix) with an attribute, \code{Afs}, for the sampling
frequency.}
}
\value{
PRH data frame with \code{rms_acc} matrix column.
}
\description{
Calculate root mean square (RMS) of acceleration in bins corresponding to
OCDR bins.
}
\examples{
prh_ocdr <- jgl_ocdr(jgl_desc(prh_expl))
acc_filt <- jgl_filtacc(Araw_expl, 10, 90)
jgl_accrms(prh_ocdr, acc_filt)

}
