% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgl_filtacc.R
\name{jgl_filtacc}
\alias{jgl_filtacc}
\title{Band pass filter acceleration}
\usage{
jgl_filtacc(A, low, high)
}
\arguments{
\item{A}{Acceleration data frame. Columns \code{Atime} (POSIXct) and
\code{A} (3-column matrix) with an attribute, \code{Afs}, for the sampling
frequency.}

\item{low, high}{Low and high cutoff frequencies for band-pass filter}
}
\value{
Same as \code{A} but band-pass filtered acceleration in acceleration
matrix (\code{A$A}) and attribute (\code{filtered}) to indicate filtering.
}
\description{
Filter the acceleration signal between low and high frequency thresholds.
}
\examples{
jgl_filtacc(Araw_expl, 10, 90)

}
