% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocdr.R
\name{jgl_ocdr}
\alias{jgl_ocdr}
\title{Orientation-corrected depth rate}
\usage{
jgl_ocdr(prh, binwidth = 1)
}
\arguments{
\item{prh}{PRH data frame with \code{desc_id} column (see
\code{\link{jgl_desc}})}

\item{binwidth}{Size of bin to calculate OCDR (in seconds)}
}
\value{
PRH data frame with \code{ocdr} column and an attribute
\code{binwidth}.
}
\description{
Uses steep descent phases to estimate known speed values.
}
\examples{
prh_expl \%>\%
  jgl_desc() \%>\%
  jgl_ocdr()

}
