% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jiggle.R
\docType{package}
\name{jiggle}
\alias{jiggle}
\alias{jiggle-package}
\title{jiggle: Estimating speed from accelerometer jiggle}
\description{
The jiggle package estimates forward speed of an animal from the amplitude
of accelerometer "jiggle". It uses an in situ calibration to regress
periods of known speed on the magnitude of the jiggle, which can be used
to predict speed throughout the deployment. See Cade et al. 2018 for more
details.
}
