% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_desc}
\alias{plot_desc}
\title{Plot descent phases within dive profile}
\usage{
plot_desc(prh)
}
\arguments{
\item{prh}{PRH data frame with \code{desc_id} column}
}
\value{
A gg object (see \code{\link[cowplot]{cowplot}})
}
\description{
Highlights descent phases used to calculate OCDR. Requires package
\code{cowplot}.
}
\examples{
plot_desc(find_desc(prh_expl))

}
