% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_filt}
\alias{plot_filt}
\title{Plot filtered acceleration}
\usage{
plot_filt(acc, prh)
}
\arguments{
\item{acc}{Filtered acceleration data frame (see \code{\link{jgl_filtacc}}).}

\item{prh}{PRH data frame.}
}
\value{
A gg object (see \code{cowplot})
}
\description{
Plots the filtered acceleration signal with depth profile for context. Note,
raw acceleration is extremely large so this function is best used with
smaller subsets.
}
\examples{
plot_filt(jgl_filtacc(Araw_expl), prh_expl)

}
