% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_ocdr}
\alias{plot_ocdr}
\title{Plot OCDR}
\usage{
plot_ocdr(prh)
}
\arguments{
\item{prh}{PRH data frame with OCDR calculated}
}
\value{
A ggplot object
}
\description{
Create a plot of orientation-corrected depth rate during descents.
}
\examples{
prh_expl \%>\%
  jgl_desc() \%>\%
  jgl_ocdr() \%>\%
  plot_ocdr()

}
