% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_rms}
\alias{plot_rms}
\title{Plot OCDR ~ jiggle}
\usage{
plot_rms(prh)
}
\arguments{
\item{prh}{PRH data frame with \code{ocdr} and \code{rms_acc} columns.}
}
\value{
A ggplot object
}
\description{
Plots OCDR against the norm of the triaxial jiggle.
}
\examples{
prh_ocdr <- jgl_ocdr(jgl_desc(prh_expl))
acc_filt <- jgl_filtacc(Araw_expl, 10, 90)
acc_rms <- jgl_accrms(prh_ocdr, acc_filt)
plot_rms(acc_rms)

}
