% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_speed}
\alias{plot_speed}
\title{Plot speed}
\usage{
plot_speed(prh, smooth = 1)
}
\arguments{
\item{prh}{PRH data frame with \code{speed} column}

\item{smooth}{Width of smoothing window in seconds}
}
\value{
A gg object (see \code{cowplot})
}
\description{
Plot speed as estimated from tag jiggle. Requires package \code{cowplot}.
}
\examples{
acc_filt <- filter_acc(Araw_expl, 10, 90)
prh_expl \%>\%
  find_desc() \%>\%
  get_ocdr() \%>\%
  get_rms_acc(acc_filt) \%>\%
  fit_model() \%>\%
  plot_speed()

}
