% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prh}
\alias{prh}
\title{Example PRH data}
\format{A list}
\source{
\href{https://doi.org/10.1242/jeb.170449}{JEB}
}
\usage{
data(prh)
}
\description{
Example data from deployment mn160727-11.
}
\references{
Cade, David E., Kelly R. Barr, John Calambokidis, Ari S.
Friedlaender, and Jeremy A. Goldbogen. “Determining Forward Speed from
Accelerometer Jiggle in Aquatic Environments.” Journal of Experimental
Biology 221, no. 2.
}
\keyword{datasets}
