% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prh_expl}
\alias{prh_expl}
\title{Example PRH data}
\format{A data frame. See \code{\link[prhdata]{prh}}.}
\source{
\href{https://doi.org/10.1242/jeb.170449}{Journal of Experimental
Biology}
}
\usage{
prh_expl
}
\description{
Example data from a CATS biologger deployed on a humpback whale (\emph{Megaptera
novaeangliae}) in Monterey Bay, CA, US on 2016-07-27.
}
\references{
Cade, David E., Kelly R. Barr, John Calambokidis, Ari S.
Friedlaender, and Jeremy A. Goldbogen. “Determining Forward Speed from
Accelerometer Jiggle in Aquatic Environments.” Journal of Experimental
Biology 221, no. 2.
}
\keyword{datasets}
