% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocdr.R
\name{find_desc}
\alias{find_desc}
\title{Find steep descent phases at the beginning of dives}
\usage{
find_desc(depth, pitch, fs, depth_thr = 5, pitch_thr = -60 * pi/180)
}
\description{
Find steep descent phases at the beginning of dives
}
